requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0

# Audio processing
ffmpeg-python>=0.2.0
pydub>=0.25.1

# Speech-to-text
faster-whisper>=0.9.0
vosk>=0.3.45

# NLP processing
spacy>=3.6.0
nltk>=3.8.1
regex>=2023.0.0
symspellpy>=6.7.7

# Video processing
moviepy>=1.0.3

# Optional - will be used in later parts
# transformers>=4.30.0
# llama-cpp-python>=0.2.0
# gpt4all>=2.0.0

# Development tools
pytest>=7.4.0
black>=23.7.0
pylint>=2.17.5
